@S-Dionis 2024-07-16
//before
stats_optimization_test.go:46: time used: 333.5098ms / 300ms
stats_optimization_test.go:47: memory used: 308Mb / 30Mb

//after
stats_optimization_test.go:46: time used: 123.0003ms / 300ms
stats_optimization_test.go:47: memory used: 15Mb / 30Mb